@@ -1,19 +1,19 @@
 name: gitjournal
 description: A Note Taking App Built on top of Git
-version: 1.88.0+10
+version: 1.89.0+10
 
 environment:
   sdk: ">=3.3.0 <4.0.0"
-  flutter: ">=3.19.6"
+  flutter: ">=3.24.2"
 
 publish_to: none
 
 dependencies:
-  android_external_storage: ^0.1.0
+  android_x_storage: ^1.0.0
   auto_size_text: ^3.0.0-nullsafety.0
   badges: ^2.0.1
   buffer: ^1.2.0
-  carousel_slider: ^4.0.0
+  carousel_slider: ^5.0.0
   collection: ^1.15.0-nullsafety.4
   community_material_icon: ^5.9.55
   connectivity_plus: ^5.0.2
@@ -23,57 +23,54 @@ dependencies:
   dart_date: ^1.1.1
   dart_git:
     git: https://github.com/GitJournal/dart-git.git
+    # path: ../gitjournal/dart-git
   device_info_plus: ^9.1.1
   dio: ^4.0.6
   dio_smart_retry: ^1.2.0
   dots_indicator: ^2.0.0
   email_validator: ^2.0.1
   equatable: ^2.0.0
-  fast_immutable_collections: ^9.1.5
+  fast_immutable_collections: ^10.2.4
   file_picker: ^5.3.2
   fimber: ^0.6.1
   fixnum: ^1.0.1
   flutter:
     sdk: flutter
   flutter_cache_manager: ^3.0.1
-  flutter_displaymode: ^0.4.1
+  flutter_displaymode: ^0.6.0
   flutter_email_sender: ^5.0.2
   flutter_emoji: ^2.4.0
   flutter_localizations:
     sdk: flutter
-  flutter_login: ^4.0.0
   flutter_markdown: ^0.7.1
-  flutter_runtime_env: ^0.0.4
   flutter_staggered_grid_view: ^0.4.1
   flutter_svg: ^2.0.7
-  flutter_tex_js: ^2.0.0
+  flutter_tex_js: ^3.0.0
   font_awesome_flutter: ^10.0.0
-  fs_shim: ^2.0.1+2
   function_types: ^0.1.0
   go_git_dart:
     git: https://github.com/GitJournal/go_git_dart.git
   git_setup:
     path: packages/git_setup
-  google_api_availability: ^3.0.1
+  google_api_availability: ^5.0.0
   google_fonts: ^2.1.0
-  grpc: ^3.0.0
   hive: ^2.2.3
   http: ^0.13.1
   icloud_documents_path:
     git: https://github.com/GitJournal/icloud_documents_path.git
   image_picker: ^0.8.4+1
-  in_app_purchase: ^3.1.1
-  in_app_purchase_android: ^0.2.3+9
-  intl: ^0.18.0
+  in_app_purchase: ^3.2.0
+  in_app_purchase_android: ^0.3.4
+  intl: ^0.19.0
   language_picker: ^0.4.1
-  launch_review: ^3.0.1
+  launch_app_store: ^1.1.2
   markdown: ^7.2.2
   meta: ^1.8.0
   mutex: ^3.0.0
   nested: ^1.0.0
   openssh_ed25519:
     git: https://github.com/GitJournal/openssh_ed25519.git
-  org_flutter: ^3.0.0
+  org_flutter: ^6.0.3
   package_info_plus: ^4.0.2
   path: ^1.8.0
   path_provider: ^2.0.11
@@ -86,21 +83,23 @@ dependencies:
   recase: ^4.0.0
   receive_sharing_intent: ^1.4.5
   sentry_flutter: ^7.8.0
-  share_plus: ^7.2.1
+  share_plus: ^10.0.2
   shared_preferences: ^2.0.5
   sprintf: ^7.0.0
   stack_trace: ^1.11.1
   synchronized: ^3.0.0
-  table_calendar: ^3.0.3
   time: ^2.0.0
   timeago: ^3.1.0
-  timeline_tile: ^2.0.0
   tuple: ^2.0.1
   universal_io: ^2.0.4
   url_launcher: ^6.0.3
   uuid: ^4.2.1
   yaml: ^3.1.0
   yaml_serializer: ^0.0.12
+  archive: ^3.6.1
+  flutter_bloc: ^8.1.6
+  bloc: ^8.1.4
+  supabase_flutter: ^2.6.0
 
 dev_dependencies:
   benchmark_harness: ^2.0.0
