@@ -123,8 +123,8 @@ class OnBoardingScreenState extends State<OnBoardingScreen> {
         child: ElevatedButton(
           key: const ValueKey("GetStarted"),
           style: ButtonStyle(
-            backgroundColor: MaterialStateProperty.all<Color>(
-                Theme.of(context).primaryColor),
+            backgroundColor:
+                WidgetStateProperty.all<Color>(Theme.of(context).primaryColor),
           ),
           onPressed: _finish,
           child: Text(
@@ -185,7 +185,7 @@ class OnBoardingBottomButton extends StatelessWidget {
       onPressed: onPressed,
       style: isDark
           ? ButtonStyle(
-              backgroundColor: MaterialStateProperty.all<Color>(
+              backgroundColor: WidgetStateProperty.all<Color>(
                 Theme.of(context).primaryColor,
               ),
             )
