@@ -1,15 +1,13 @@
+// ignore_for_file: avoid_print
+
 /*
  * SPDX-FileCopyrightText: 2019-2021 Vishesh Handa <me@vhanda.in>
  *
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
 
-import 'package:email_validator/email_validator.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_login/flutter_login.dart';
-import 'package:gitjournal/account/account_screen.dart';
-import 'package:gitjournal/l10n.dart';
-import 'package:gitjournal/logger/logger.dart';
+import 'package:supabase_flutter/supabase_flutter.dart';
 
 class LoginPage extends StatefulWidget {
   static const routePath = '/login';
@@ -26,222 +24,63 @@ class _LoginPageState extends State<LoginPage> {
   @override
   void initState() {
     super.initState();
+
+    _initAsync();
+  }
+
+  Future<void> _initAsync() async {
+    await Supabase.initialize(
+      url: 'https://kokulrmhlfxdwuvcmblj.supabase.co',
+      anonKey: '<key>',
+    );
   }
 
   @override
   Widget build(BuildContext context) {
-    var login = FlutterLogin(
-      title: 'GitJournal',
-      logo: 'assets/icon/icon.png',
-      navigateBackAfterRecovery: false,
-      loginAfterSignUp: false,
-      // hideForgotPasswordButton: true,
-      // hideSignUpButton: true,
-      // disableCustomPageTransformer: true,
-      // messages: LoginMessages(
-      //   userHint: 'User',
-      //   passwordHint: 'Pass',
-      //   confirmPasswordHint: 'Confirm',
-      //   loginButton: 'LOG IN',
-      //   signupButton: 'REGISTER',
-      //   forgotPasswordButton: 'Forgot huh?',
-      //   recoverPasswordButton: 'HELP ME',
-      //   goBackButton: 'GO BACK',
-      //   confirmPasswordError: 'Not match!',
-      //   recoverPasswordIntro: 'Don\'t feel bad. Happens all the time.',
-      //   recoverPasswordDescription: 'Lorem Ipsum is simply dummy text of the printing and typesetting industry',
-      //   recoverPasswordSuccess: 'Password rescued successfully',
-      //   flushbarTitleError: 'Oh no!',
-      //   flushbarTitleSuccess: 'Succes!',
-      //   providersTitle: 'login with'
-      // ),
-      // theme: LoginTheme(
-      //   primaryColor: Colors.teal,
-      //   accentColor: Colors.yellow,
-      //   errorColor: Colors.deepOrange,
-      //   pageColorLight: Colors.indigo.shade300,
-      //   pageColorDark: Colors.indigo.shade500,
-      //   logoWidth: 0.80,
-      //   titleStyle: TextStyle(
-      //     color: Colors.greenAccent,
-      //     fontFamily: 'Quicksand',
-      //     letterSpacing: 4,
-      //   ),
-      //   // beforeHeroFontSize: 50,
-      //   // afterHeroFontSize: 20,
-      //   bodyStyle: TextStyle(
-      //     fontStyle: FontStyle.italic,
-      //     decoration: TextDecoration.underline,
-      //   ),
-      //   textFieldStyle: TextStyle(
-      //     color: Colors.orange,
-      //     shadows: [Shadow(color: Colors.yellow, blurRadius: 2)],
-      //   ),
-      //   buttonStyle: TextStyle(
-      //     fontWeight: FontWeight.w800,
-      //     color: Colors.yellow,
-      //   ),
-      //   cardTheme: CardTheme(
-      //     color: Colors.yellow.shade100,
-      //     elevation: 5,
-      //     margin: EdgeInsets.only(top: 15),
-      //     shape: ContinuousRectangleBorder(
-      //         borderRadius: BorderRadius.circular(100.0)),
-      //   ),
-      //   inputTheme: InputDecorationTheme(
-      //     filled: true,
-      //     fillColor: Colors.purple.withOpacity(.1),
-      //     contentPadding: EdgeInsets.zero,
-      //     errorStyle: TextStyle(
-      //       backgroundColor: Colors.orange,
-      //       color: Colors.white,
-      //     ),
-      //     labelStyle: TextStyle(fontSize: 12),
-      //     enabledBorder: UnderlineInputBorder(
-      //       borderSide: BorderSide(color: Colors.blue.shade700, width: 4),
-      //       borderRadius: inputBorder,
-      //     ),
-      //     focusedBorder: UnderlineInputBorder(
-      //       borderSide: BorderSide(color: Colors.blue.shade400, width: 5),
-      //       borderRadius: inputBorder,
-      //     ),
-      //     errorBorder: UnderlineInputBorder(
-      //       borderSide: BorderSide(color: Colors.red.shade700, width: 7),
-      //       borderRadius: inputBorder,
-      //     ),
-      //     focusedErrorBorder: UnderlineInputBorder(
-      //       borderSide: BorderSide(color: Colors.red.shade400, width: 8),
-      //       borderRadius: inputBorder,
-      //     ),
-      //     disabledBorder: UnderlineInputBorder(
-      //       borderSide: BorderSide(color: Colors.grey, width: 5),
-      //       borderRadius: inputBorder,
-      //     ),
-      //   ),
-      //   buttonTheme: LoginButtonTheme(
-      //     splashColor: Colors.purple,
-      //     backgroundColor: Colors.pinkAccent,
-      //     highlightColor: Colors.lightGreen,
-      //     elevation: 9.0,
-      //     highlightElevation: 6.0,
-      //     shape: BeveledRectangleBorder(
-      //       borderRadius: BorderRadius.circular(10),
-      //     ),
-      //     // shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(5)),
-      //     // shape: CircleBorder(side: BorderSide(color: Colors.green)),
-      //     // shape: ContinuousRectangleBorder(borderRadius: BorderRadius.circular(55.0)),
-      //   ),
-      // ),
-      termsOfService: [
-        TermOfService(
-          id: 'newsletter',
-          mandatory: false,
-          text: 'Subscribe to the GitJournal Newsletter',
-        ),
-      ],
-      userValidator: (value) {
-        if (value == null || value.isEmpty) {
-          return "Email is empty";
-        }
-        if (!EmailValidator.validate(value)) {
-          return context.loc.settingsEmailValidatorInvalid;
-        }
-        return null;
-      },
-      passwordValidator: (value) {
-        if (value == null || value.isEmpty) {
-          return 'Password is empty';
-        }
-        return null;
-      },
-      onLogin: _login,
-      onSignup: _signup,
-      onSubmitAnimationCompleted: () {
-        Navigator.of(context).pushReplacement(_FadePageRoute(
-          builder: (context) => const AccountScreen(),
-        ));
-      },
-      onRecoverPassword: (name) async {
-        Log.i('Recover password info');
-        Log.i('Name: $name');
-        return "";
-        // return _recoverPassword(name);
-        // Show new password dialog
-      },
-      disableCustomPageTransformer: true,
-      // showDebugButtons: true,
-    );
-
     return Scaffold(
       appBar: AppBar(
         backgroundColor: Theme.of(context).primaryColor,
         elevation: 0.0,
       ),
-      body: login,
-    );
-  }
-
-  Future<String?> _signup(SignupData signupData) async {
-    Log.i('Signup info');
-    Log.i('Name: ${signupData.name}');
-    Log.i('Password: ${signupData.password}');
-
-    return "";
-  }
-
-  Future<String?> _login(LoginData loginData) async {
-    return "";
-  }
-}
-
-class FormBackButton extends StatelessWidget {
-  const FormBackButton({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return InkWell(
-      onTap: () {
-        Navigator.pop(context);
-      },
-      child: Container(
-        padding: const EdgeInsets.symmetric(horizontal: 10),
-        child: const Row(
-          children: <Widget>[
-            Icon(Icons.keyboard_arrow_left, color: Colors.black),
-            Text(
-              'Back',
-              style: TextStyle(fontSize: 12, fontWeight: FontWeight.w500),
-            )
-          ],
-        ),
+      body: Column(
+        children: [
+          TextButton(
+            onPressed: _login,
+            child: const Text('Login'),
+          ),
+        ],
       ),
     );
   }
-}
 
-class _FadePageRoute<T> extends MaterialPageRoute<T> {
-  _FadePageRoute({
-    required super.builder,
-    super.settings,
-  });
-
-  @override
-  Duration get transitionDuration => const Duration(milliseconds: 600);
-
-  @override
-  Widget buildTransitions(
-    BuildContext context,
-    Animation<double> animation,
-    Animation<double> secondaryAnimation,
-    Widget child,
-  ) {
-    if (settings.name == LoginPage.routePath) {
-      return child;
-    }
-
-    return FadeTransition(
-      opacity: animation,
-      child: child,
+  Future<String?> _login() async {
+    final supabase = Supabase.instance.client;
+    final AuthResponse res = await supabase.auth.signInWithPassword(
+      email: 'test@gitjournal.io',
+      password: '<pass>',
     );
+    final session = res.session;
+    final user = res.user;
+
+    print("User: $user");
+    print("Session: $session");
+
+    final qr = await supabase.from('analytics_device_info').select();
+    print(qr);
+    return "";
   }
 }
+
+/*
+
+Next Steps:
+
+1. Create a simple table about user_info which has the bool 'pro'
+2. Add RLS to that table so only the user can see read own data
+3. Add RLS to the other tables (make sure no one can read those)
+4. Figure out how the Supabase session data is saved!
+5. Create a very simple login/password screen, and just allow the user to login
+6. Direct the user to the account_screen or login_screen depending if they are logged in
+7. Add a logout button to the account_screen
+
+ */
