@@ -1,51 +1,27 @@
-// SPDX-FileCopyrightText: 2021 Vishesh Handa <me@vhanda.in>
 //
-// SPDX-License-Identifier: AGPL-3.0-or-later
-
-///
 //  Generated code. Do not modify.
 //  source: analytics.proto
 //
 // @dart = 2.12
-// ignore_for_file: annotate_overrides,camel_case_types,unnecessary_const,non_constant_identifier_names,library_prefixes,unused_import,unused_shown_name,return_of_invalid_type,unnecessary_this,prefer_final_fields
 
-// ignore_for_file: UNDEFINED_SHOWN_NAME
+// ignore_for_file: annotate_overrides, camel_case_types, comment_references
+// ignore_for_file: constant_identifier_names, library_prefixes
+// ignore_for_file: non_constant_identifier_names, prefer_final_fields
+// ignore_for_file: unnecessary_import, unnecessary_this, unused_import
 
 import 'dart:core' as $core;
 
 import 'package:protobuf/protobuf.dart' as $pb;
 
 class Platform extends $pb.ProtobufEnum {
-  static const Platform android = Platform._(
-      0,
-      const $core.bool.fromEnvironment('protobuf.omit_enum_names')
-          ? ''
-          : 'android');
-  static const Platform ios = Platform._(
-      1,
-      const $core.bool.fromEnvironment('protobuf.omit_enum_names')
-          ? ''
-          : 'ios');
-  static const Platform linux = Platform._(
-      2,
-      const $core.bool.fromEnvironment('protobuf.omit_enum_names')
-          ? ''
-          : 'linux');
-  static const Platform macos = Platform._(
-      3,
-      const $core.bool.fromEnvironment('protobuf.omit_enum_names')
-          ? ''
-          : 'macos');
-  static const Platform windows = Platform._(
-      4,
-      const $core.bool.fromEnvironment('protobuf.omit_enum_names')
-          ? ''
-          : 'windows');
-  static const Platform web = Platform._(
-      5,
-      const $core.bool.fromEnvironment('protobuf.omit_enum_names')
-          ? ''
-          : 'web');
+  static const Platform android =
+      Platform._(0, _omitEnumNames ? '' : 'android');
+  static const Platform ios = Platform._(1, _omitEnumNames ? '' : 'ios');
+  static const Platform linux = Platform._(2, _omitEnumNames ? '' : 'linux');
+  static const Platform macos = Platform._(3, _omitEnumNames ? '' : 'macos');
+  static const Platform windows =
+      Platform._(4, _omitEnumNames ? '' : 'windows');
+  static const Platform web = Platform._(5, _omitEnumNames ? '' : 'web');
 
   static const $core.List<Platform> values = <Platform>[
     android,
@@ -64,46 +40,22 @@ class Platform extends $pb.ProtobufEnum {
 }
 
 class BrowserName extends $pb.ProtobufEnum {
-  static const BrowserName unknown = BrowserName._(
-      0,
-      const $core.bool.fromEnvironment('protobuf.omit_enum_names')
-          ? ''
-          : 'unknown');
-  static const BrowserName firefox = BrowserName._(
-      1,
-      const $core.bool.fromEnvironment('protobuf.omit_enum_names')
-          ? ''
-          : 'firefox');
-  static const BrowserName samsungInternet = BrowserName._(
-      2,
-      const $core.bool.fromEnvironment('protobuf.omit_enum_names')
-          ? ''
-          : 'samsungInternet');
-  static const BrowserName opera = BrowserName._(
-      3,
-      const $core.bool.fromEnvironment('protobuf.omit_enum_names')
-          ? ''
-          : 'opera');
-  static const BrowserName msie = BrowserName._(
-      4,
-      const $core.bool.fromEnvironment('protobuf.omit_enum_names')
-          ? ''
-          : 'msie');
-  static const BrowserName edge = BrowserName._(
-      5,
-      const $core.bool.fromEnvironment('protobuf.omit_enum_names')
-          ? ''
-          : 'edge');
-  static const BrowserName chrome = BrowserName._(
-      6,
-      const $core.bool.fromEnvironment('protobuf.omit_enum_names')
-          ? ''
-          : 'chrome');
-  static const BrowserName safari = BrowserName._(
-      7,
-      const $core.bool.fromEnvironment('protobuf.omit_enum_names')
-          ? ''
-          : 'safari');
+  static const BrowserName unknown =
+      BrowserName._(0, _omitEnumNames ? '' : 'unknown');
+  static const BrowserName firefox =
+      BrowserName._(1, _omitEnumNames ? '' : 'firefox');
+  static const BrowserName samsungInternet =
+      BrowserName._(2, _omitEnumNames ? '' : 'samsungInternet');
+  static const BrowserName opera =
+      BrowserName._(3, _omitEnumNames ? '' : 'opera');
+  static const BrowserName msie =
+      BrowserName._(4, _omitEnumNames ? '' : 'msie');
+  static const BrowserName edge =
+      BrowserName._(5, _omitEnumNames ? '' : 'edge');
+  static const BrowserName chrome =
+      BrowserName._(6, _omitEnumNames ? '' : 'chrome');
+  static const BrowserName safari =
+      BrowserName._(7, _omitEnumNames ? '' : 'safari');
 
   static const $core.List<BrowserName> values = <BrowserName>[
     unknown,
@@ -122,3 +74,5 @@ class BrowserName extends $pb.ProtobufEnum {
 
   const BrowserName._($core.int v, $core.String n) : super(v, n);
 }
+
+const _omitEnumNames = $core.bool.fromEnvironment('protobuf.omit_enum_names');
