@@ -11,13 +11,12 @@ import 'package:gitjournal/account/account_screen.dart';
 import 'package:gitjournal/account/login_screen.dart';
 import 'package:gitjournal/core/markdown/md_yaml_doc_codec.dart';
 import 'package:gitjournal/editors/note_editor.dart';
-import 'package:gitjournal/history/history_screen.dart';
+import 'package:gitjournal/folder_listing/view/folder_listing.dart';
 import 'package:gitjournal/iap/purchase_screen.dart';
 import 'package:gitjournal/iap/purchase_thankyou_screen.dart';
 import 'package:gitjournal/logger/logger.dart';
 import 'package:gitjournal/repository.dart';
 import 'package:gitjournal/screens/error_screen.dart';
-import 'package:gitjournal/screens/folder_listing.dart';
 import 'package:gitjournal/screens/home_screen.dart';
 import 'package:gitjournal/screens/onboarding_screens.dart';
 import 'package:gitjournal/screens/tag_listing.dart';
@@ -41,7 +40,6 @@ class AppRoute {
     GitHostSetupScreen.routePath,
     PurchaseScreen.routePath,
     PurchaseThankYouScreen.routePath,
-    HistoryScreen.routePath,
     ErrorScreen.routePath,
   ];
 }
@@ -140,8 +138,6 @@ class AppRouter {
         return PurchaseScreen();
       case PurchaseThankYouScreen.routePath:
         return PurchaseThankYouScreen();
-      case HistoryScreen.routePath:
-        return const HistoryScreen();
       case ErrorScreen.routePath:
         return const ErrorScreen();
     }
