@@ -1,79 +1,87 @@
-// SPDX-FileCopyrightText: 2021 Vishesh Handa <me@vhanda.in>
 //
-// SPDX-License-Identifier: AGPL-3.0-or-later
-
-///
 //  Generated code. Do not modify.
 //  source: core.proto
 //
 // @dart = 2.12
-// ignore_for_file: annotate_overrides,camel_case_types,unnecessary_const,non_constant_identifier_names,library_prefixes,unused_import,unused_shown_name,return_of_invalid_type,unnecessary_this,prefer_final_fields,deprecated_member_use_from_same_package
+
+// ignore_for_file: annotate_overrides, camel_case_types, comment_references
+// ignore_for_file: constant_identifier_names, library_prefixes
+// ignore_for_file: non_constant_identifier_names, prefer_final_fields
+// ignore_for_file: unnecessary_import, unnecessary_this, unused_import
 
 import 'dart:convert' as $convert;
 import 'dart:core' as $core;
 import 'dart:typed_data' as $typed_data;
 
 @$core.Deprecated('Use noteFileFormatDescriptor instead')
-const NoteFileFormat$json = const {
+const NoteFileFormat$json = {
   '1': 'NoteFileFormat',
-  '2': const [
-    const {'1': 'Markdown', '2': 0},
-    const {'1': 'OrgMode', '2': 1},
-    const {'1': 'Txt', '2': 2},
+  '2': [
+    {'1': 'Markdown', '2': 0},
+    {'1': 'OrgMode', '2': 1},
+    {'1': 'Txt', '2': 2},
   ],
 };
 
 /// Descriptor for `NoteFileFormat`. Decode as a `google.protobuf.EnumDescriptorProto`.
 final $typed_data.Uint8List noteFileFormatDescriptor = $convert.base64Decode(
     'Cg5Ob3RlRmlsZUZvcm1hdBIMCghNYXJrZG93bhAAEgsKB09yZ01vZGUQARIHCgNUeHQQAg==');
+
 @$core.Deprecated('Use noteTypeDescriptor instead')
-const NoteType$json = const {
+const NoteType$json = {
   '1': 'NoteType',
-  '2': const [
-    const {'1': 'Unknown', '2': 0},
-    const {'1': 'Checklist', '2': 1},
-    const {'1': 'Journal', '2': 2},
-    const {'1': 'Org', '2': 3},
+  '2': [
+    {'1': 'Unknown', '2': 0},
+    {'1': 'Checklist', '2': 1},
+    {'1': 'Journal', '2': 2},
+    {'1': 'Org', '2': 3},
   ],
 };
 
 /// Descriptor for `NoteType`. Decode as a `google.protobuf.EnumDescriptorProto`.
 final $typed_data.Uint8List noteTypeDescriptor = $convert.base64Decode(
-    'CghOb3RlVHlwZRILCgdVbmtub3duEAASDQoJQ2hlY2tsaXN0EAESCwoHSm91cm5hbBACEgcKA09yZxAD');
+    'CghOb3RlVHlwZRILCgdVbmtub3duEAASDQoJQ2hlY2tsaXN0EAESCwoHSm91cm5hbBACEgcKA0'
+    '9yZxAD');
+
 @$core.Deprecated('Use unixTimestampMagnitudeDescriptor instead')
-const UnixTimestampMagnitude$json = const {
+const UnixTimestampMagnitude$json = {
   '1': 'UnixTimestampMagnitude',
-  '2': const [
-    const {'1': 'Seconds', '2': 0},
-    const {'1': 'Milliseconds', '2': 1},
+  '2': [
+    {'1': 'Seconds', '2': 0},
+    {'1': 'Milliseconds', '2': 1},
   ],
 };
 
 /// Descriptor for `UnixTimestampMagnitude`. Decode as a `google.protobuf.EnumDescriptorProto`.
 final $typed_data.Uint8List unixTimestampMagnitudeDescriptor =
     $convert.base64Decode(
-        'ChZVbml4VGltZXN0YW1wTWFnbml0dWRlEgsKB1NlY29uZHMQABIQCgxNaWxsaXNlY29uZHMQAQ==');
+        'ChZVbml4VGltZXN0YW1wTWFnbml0dWRlEgsKB1NlY29uZHMQABIQCgxNaWxsaXNlY29uZHMQAQ'
+        '==');
+
 @$core.Deprecated('Use dateFormatDescriptor instead')
-const DateFormat$json = const {
+const DateFormat$json = {
   '1': 'DateFormat',
-  '2': const [
-    const {'1': 'Iso8601', '2': 0},
-    const {'1': 'UnixTimeStamp', '2': 1},
-    const {'1': 'None', '2': 2},
+  '2': [
+    {'1': 'Iso8601', '2': 0},
+    {'1': 'UnixTimeStamp', '2': 1},
+    {'1': 'None', '2': 2},
+    {'1': 'YearMonthDay', '2': 3},
   ],
 };
 
 /// Descriptor for `DateFormat`. Decode as a `google.protobuf.EnumDescriptorProto`.
 final $typed_data.Uint8List dateFormatDescriptor = $convert.base64Decode(
-    'CgpEYXRlRm9ybWF0EgsKB0lzbzg2MDEQABIRCg1Vbml4VGltZVN0YW1wEAESCAoETm9uZRAC');
+    'CgpEYXRlRm9ybWF0EgsKB0lzbzg2MDEQABIRCg1Vbml4VGltZVN0YW1wEAESCAoETm9uZRACEh'
+    'AKDFllYXJNb250aERheRAD');
+
 @$core.Deprecated('Use fileDescriptor instead')
-const File$json = const {
+const File$json = {
   '1': 'File',
-  '2': const [
-    const {'1': 'repoPath', '3': 1, '4': 1, '5': 9, '10': 'repoPath'},
-    const {'1': 'hash', '3': 2, '4': 1, '5': 12, '10': 'hash'},
-    const {'1': 'filePath', '3': 3, '4': 1, '5': 9, '10': 'filePath'},
-    const {
+  '2': [
+    {'1': 'repoPath', '3': 1, '4': 1, '5': 9, '10': 'repoPath'},
+    {'1': 'hash', '3': 2, '4': 1, '5': 12, '10': 'hash'},
+    {'1': 'filePath', '3': 3, '4': 1, '5': 9, '10': 'filePath'},
+    {
       '1': 'modified',
       '3': 4,
       '4': 1,
@@ -81,7 +89,7 @@ const File$json = const {
       '6': '.gitjournal.DateTimeAnyTz',
       '10': 'modified'
     },
-    const {
+    {
       '1': 'created',
       '3': 5,
       '4': 1,
@@ -89,7 +97,7 @@ const File$json = const {
       '6': '.gitjournal.DateTimeAnyTz',
       '10': 'created'
     },
-    const {
+    {
       '1': 'fileLastModified',
       '3': 6,
       '4': 1,
@@ -102,12 +110,17 @@ const File$json = const {
 
 /// Descriptor for `File`. Decode as a `google.protobuf.DescriptorProto`.
 final $typed_data.Uint8List fileDescriptor = $convert.base64Decode(
-    'CgRGaWxlEhoKCHJlcG9QYXRoGAEgASgJUghyZXBvUGF0aBISCgRoYXNoGAIgASgMUgRoYXNoEhoKCGZpbGVQYXRoGAMgASgJUghmaWxlUGF0aBI1Cghtb2RpZmllZBgEIAEoCzIZLmdpdGpvdXJuYWwuRGF0ZVRpbWVBbnlUelIIbW9kaWZpZWQSMwoHY3JlYXRlZBgFIAEoCzIZLmdpdGpvdXJuYWwuRGF0ZVRpbWVBbnlUelIHY3JlYXRlZBJFChBmaWxlTGFzdE1vZGlmaWVkGAYgASgLMhkuZ2l0am91cm5hbC5EYXRlVGltZUFueVR6UhBmaWxlTGFzdE1vZGlmaWVk');
+    'CgRGaWxlEhoKCHJlcG9QYXRoGAEgASgJUghyZXBvUGF0aBISCgRoYXNoGAIgASgMUgRoYXNoEh'
+    'oKCGZpbGVQYXRoGAMgASgJUghmaWxlUGF0aBI1Cghtb2RpZmllZBgEIAEoCzIZLmdpdGpvdXJu'
+    'YWwuRGF0ZVRpbWVBbnlUelIIbW9kaWZpZWQSMwoHY3JlYXRlZBgFIAEoCzIZLmdpdGpvdXJuYW'
+    'wuRGF0ZVRpbWVBbnlUelIHY3JlYXRlZBJFChBmaWxlTGFzdE1vZGlmaWVkGAYgASgLMhkuZ2l0'
+    'am91cm5hbC5EYXRlVGltZUFueVR6UhBmaWxlTGFzdE1vZGlmaWVk');
+
 @$core.Deprecated('Use noteDescriptor instead')
-const Note$json = const {
+const Note$json = {
   '1': 'Note',
-  '2': const [
-    const {
+  '2': [
+    {
       '1': 'file',
       '3': 1,
       '4': 1,
@@ -115,9 +128,9 @@ const Note$json = const {
       '6': '.gitjournal.File',
       '10': 'file'
     },
-    const {'1': 'title', '3': 2, '4': 1, '5': 9, '10': 'title'},
-    const {'1': 'body', '3': 3, '4': 1, '5': 9, '10': 'body'},
-    const {
+    {'1': 'title', '3': 2, '4': 1, '5': 9, '10': 'title'},
+    {'1': 'body', '3': 3, '4': 1, '5': 9, '10': 'body'},
+    {
       '1': 'type',
       '3': 4,
       '4': 1,
@@ -125,8 +138,8 @@ const Note$json = const {
       '6': '.gitjournal.NoteType',
       '10': 'type'
     },
-    const {'1': 'tags', '3': 5, '4': 3, '5': 9, '10': 'tags'},
-    const {
+    {'1': 'tags', '3': 5, '4': 3, '5': 9, '10': 'tags'},
+    {
       '1': 'extraProps',
       '3': 6,
       '4': 3,
@@ -134,7 +147,7 @@ const Note$json = const {
       '6': '.gitjournal.Note.ExtraPropsEntry',
       '10': 'extraProps'
     },
-    const {
+    {
       '1': 'fileFormat',
       '3': 7,
       '4': 1,
@@ -142,8 +155,8 @@ const Note$json = const {
       '6': '.gitjournal.NoteFileFormat',
       '10': 'fileFormat'
     },
-    const {'1': 'propsList', '3': 13, '4': 3, '5': 9, '10': 'propsList'},
-    const {
+    {'1': 'propsList', '3': 13, '4': 3, '5': 9, '10': 'propsList'},
+    {
       '1': 'modified',
       '3': 10,
       '4': 1,
@@ -151,7 +164,7 @@ const Note$json = const {
       '6': '.gitjournal.DateTimeAnyTz',
       '10': 'modified'
     },
-    const {
+    {
       '1': 'created',
       '3': 11,
       '4': 1,
@@ -159,7 +172,7 @@ const Note$json = const {
       '6': '.gitjournal.DateTimeAnyTz',
       '10': 'created'
     },
-    const {
+    {
       '1': 'serializerSettings',
       '3': 12,
       '4': 1,
@@ -168,15 +181,15 @@ const Note$json = const {
       '10': 'serializerSettings'
     },
   ],
-  '3': const [Note_ExtraPropsEntry$json],
+  '3': [Note_ExtraPropsEntry$json],
 };
 
 @$core.Deprecated('Use noteDescriptor instead')
-const Note_ExtraPropsEntry$json = const {
+const Note_ExtraPropsEntry$json = {
   '1': 'ExtraPropsEntry',
-  '2': const [
-    const {'1': 'key', '3': 1, '4': 1, '5': 9, '10': 'key'},
-    const {
+  '2': [
+    {'1': 'key', '3': 1, '4': 1, '5': 9, '10': 'key'},
+    {
       '1': 'value',
       '3': 2,
       '4': 1,
@@ -185,17 +198,28 @@ const Note_ExtraPropsEntry$json = const {
       '10': 'value'
     },
   ],
-  '7': const {'7': true},
+  '7': {'7': true},
 };
 
 /// Descriptor for `Note`. Decode as a `google.protobuf.DescriptorProto`.
 final $typed_data.Uint8List noteDescriptor = $convert.base64Decode(
-    'CgROb3RlEiQKBGZpbGUYASABKAsyEC5naXRqb3VybmFsLkZpbGVSBGZpbGUSFAoFdGl0bGUYAiABKAlSBXRpdGxlEhIKBGJvZHkYAyABKAlSBGJvZHkSKAoEdHlwZRgEIAEoDjIULmdpdGpvdXJuYWwuTm90ZVR5cGVSBHR5cGUSEgoEdGFncxgFIAMoCVIEdGFncxJACgpleHRyYVByb3BzGAYgAygLMiAuZ2l0am91cm5hbC5Ob3RlLkV4dHJhUHJvcHNFbnRyeVIKZXh0cmFQcm9wcxI6CgpmaWxlRm9ybWF0GAcgASgOMhouZ2l0am91cm5hbC5Ob3RlRmlsZUZvcm1hdFIKZmlsZUZvcm1hdBIcCglwcm9wc0xpc3QYDSADKAlSCXByb3BzTGlzdBI1Cghtb2RpZmllZBgKIAEoCzIZLmdpdGpvdXJuYWwuRGF0ZVRpbWVBbnlUelIIbW9kaWZpZWQSMwoHY3JlYXRlZBgLIAEoCzIZLmdpdGpvdXJuYWwuRGF0ZVRpbWVBbnlUelIHY3JlYXRlZBJVChJzZXJpYWxpemVyU2V0dGluZ3MYDCABKAsyJS5naXRqb3VybmFsLk5vdGVTZXJpYWxpemF0aW9uU2V0dGluZ3NSEnNlcmlhbGl6ZXJTZXR0aW5ncxpQCg9FeHRyYVByb3BzRW50cnkSEAoDa2V5GAEgASgJUgNrZXkSJwoFdmFsdWUYAiABKAsyES5naXRqb3VybmFsLlVuaW9uUgV2YWx1ZToCOAE=');
+    'CgROb3RlEiQKBGZpbGUYASABKAsyEC5naXRqb3VybmFsLkZpbGVSBGZpbGUSFAoFdGl0bGUYAi'
+    'ABKAlSBXRpdGxlEhIKBGJvZHkYAyABKAlSBGJvZHkSKAoEdHlwZRgEIAEoDjIULmdpdGpvdXJu'
+    'YWwuTm90ZVR5cGVSBHR5cGUSEgoEdGFncxgFIAMoCVIEdGFncxJACgpleHRyYVByb3BzGAYgAy'
+    'gLMiAuZ2l0am91cm5hbC5Ob3RlLkV4dHJhUHJvcHNFbnRyeVIKZXh0cmFQcm9wcxI6CgpmaWxl'
+    'Rm9ybWF0GAcgASgOMhouZ2l0am91cm5hbC5Ob3RlRmlsZUZvcm1hdFIKZmlsZUZvcm1hdBIcCg'
+    'lwcm9wc0xpc3QYDSADKAlSCXByb3BzTGlzdBI1Cghtb2RpZmllZBgKIAEoCzIZLmdpdGpvdXJu'
+    'YWwuRGF0ZVRpbWVBbnlUelIIbW9kaWZpZWQSMwoHY3JlYXRlZBgLIAEoCzIZLmdpdGpvdXJuYW'
+    'wuRGF0ZVRpbWVBbnlUelIHY3JlYXRlZBJVChJzZXJpYWxpemVyU2V0dGluZ3MYDCABKAsyJS5n'
+    'aXRqb3VybmFsLk5vdGVTZXJpYWxpemF0aW9uU2V0dGluZ3NSEnNlcmlhbGl6ZXJTZXR0aW5ncx'
+    'pQCg9FeHRyYVByb3BzRW50cnkSEAoDa2V5GAEgASgJUgNrZXkSJwoFdmFsdWUYAiABKAsyES5n'
+    'aXRqb3VybmFsLlVuaW9uUgV2YWx1ZToCOAE=');
+
 @$core.Deprecated('Use noteListDescriptor instead')
-const NoteList$json = const {
+const NoteList$json = {
   '1': 'NoteList',
-  '2': const [
-    const {
+  '2': [
+    {
       '1': 'notes',
       '3': 1,
       '4': 3,
@@ -209,12 +233,13 @@ const NoteList$json = const {
 /// Descriptor for `NoteList`. Decode as a `google.protobuf.DescriptorProto`.
 final $typed_data.Uint8List noteListDescriptor = $convert.base64Decode(
     'CghOb3RlTGlzdBImCgVub3RlcxgBIAMoCzIQLmdpdGpvdXJuYWwuTm90ZVIFbm90ZXM=');
+
 @$core.Deprecated('Use mdYamlDocDescriptor instead')
-const MdYamlDoc$json = const {
+const MdYamlDoc$json = {
   '1': 'MdYamlDoc',
-  '2': const [
-    const {'1': 'body', '3': 1, '4': 1, '5': 9, '10': 'body'},
-    const {
+  '2': [
+    {'1': 'body', '3': 1, '4': 1, '5': 9, '10': 'body'},
+    {
       '1': 'map',
       '3': 2,
       '4': 3,
@@ -223,15 +248,15 @@ const MdYamlDoc$json = const {
       '10': 'map'
     },
   ],
-  '3': const [MdYamlDoc_MapEntry$json],
+  '3': [MdYamlDoc_MapEntry$json],
 };
 
 @$core.Deprecated('Use mdYamlDocDescriptor instead')
-const MdYamlDoc_MapEntry$json = const {
+const MdYamlDoc_MapEntry$json = {
   '1': 'MapEntry',
-  '2': const [
-    const {'1': 'key', '3': 1, '4': 1, '5': 9, '10': 'key'},
-    const {
+  '2': [
+    {'1': 'key', '3': 1, '4': 1, '5': 9, '10': 'key'},
+    {
       '1': 'value',
       '3': 2,
       '4': 1,
@@ -240,33 +265,22 @@ const MdYamlDoc_MapEntry$json = const {
       '10': 'value'
     },
   ],
-  '7': const {'7': true},
+  '7': {'7': true},
 };
 
 /// Descriptor for `MdYamlDoc`. Decode as a `google.protobuf.DescriptorProto`.
 final $typed_data.Uint8List mdYamlDocDescriptor = $convert.base64Decode(
-    'CglNZFlhbWxEb2MSEgoEYm9keRgBIAEoCVIEYm9keRIwCgNtYXAYAiADKAsyHi5naXRqb3VybmFsLk1kWWFtbERvYy5NYXBFbnRyeVIDbWFwGkkKCE1hcEVudHJ5EhAKA2tleRgBIAEoCVIDa2V5EicKBXZhbHVlGAIgASgLMhEuZ2l0am91cm5hbC5VbmlvblIFdmFsdWU6AjgB');
+    'CglNZFlhbWxEb2MSEgoEYm9keRgBIAEoCVIEYm9keRIwCgNtYXAYAiADKAsyHi5naXRqb3Vybm'
+    'FsLk1kWWFtbERvYy5NYXBFbnRyeVIDbWFwGkkKCE1hcEVudHJ5EhAKA2tleRgBIAEoCVIDa2V5'
+    'EicKBXZhbHVlGAIgASgLMhEuZ2l0am91cm5hbC5VbmlvblIFdmFsdWU6AjgB');
+
 @$core.Deprecated('Use unionDescriptor instead')
-const Union$json = const {
+const Union$json = {
   '1': 'Union',
-  '2': const [
-    const {
-      '1': 'booleanValue',
-      '3': 1,
-      '4': 1,
-      '5': 8,
-      '9': 0,
-      '10': 'booleanValue'
-    },
-    const {
-      '1': 'stringValue',
-      '3': 2,
-      '4': 1,
-      '5': 9,
-      '9': 0,
-      '10': 'stringValue'
-    },
-    const {
+  '2': [
+    {'1': 'booleanValue', '3': 1, '4': 1, '5': 8, '9': 0, '10': 'booleanValue'},
+    {'1': 'stringValue', '3': 2, '4': 1, '5': 9, '9': 0, '10': 'stringValue'},
+    {
       '1': 'dateValue',
       '3': 3,
       '4': 1,
@@ -275,9 +289,9 @@ const Union$json = const {
       '9': 0,
       '10': 'dateValue'
     },
-    const {'1': 'intValue', '3': 4, '4': 1, '5': 3, '9': 0, '10': 'intValue'},
-    const {'1': 'isNull', '3': 7, '4': 1, '5': 8, '9': 0, '10': 'isNull'},
-    const {
+    {'1': 'intValue', '3': 4, '4': 1, '5': 3, '9': 0, '10': 'intValue'},
+    {'1': 'isNull', '3': 7, '4': 1, '5': 8, '9': 0, '10': 'isNull'},
+    {
       '1': 'listValue',
       '3': 5,
       '4': 3,
@@ -285,7 +299,7 @@ const Union$json = const {
       '6': '.gitjournal.Union',
       '10': 'listValue'
     },
-    const {
+    {
       '1': 'mapValue',
       '3': 6,
       '4': 3,
@@ -294,18 +308,18 @@ const Union$json = const {
       '10': 'mapValue'
     },
   ],
-  '3': const [Union_MapValueEntry$json],
-  '8': const [
-    const {'1': 'UnionOneof'},
+  '3': [Union_MapValueEntry$json],
+  '8': [
+    {'1': 'UnionOneof'},
   ],
 };
 
 @$core.Deprecated('Use unionDescriptor instead')
-const Union_MapValueEntry$json = const {
+const Union_MapValueEntry$json = {
   '1': 'MapValueEntry',
-  '2': const [
-    const {'1': 'key', '3': 1, '4': 1, '5': 9, '10': 'key'},
-    const {
+  '2': [
+    {'1': 'key', '3': 1, '4': 1, '5': 9, '10': 'key'},
+    {
       '1': 'value',
       '3': 2,
       '4': 1,
@@ -314,37 +328,47 @@ const Union_MapValueEntry$json = const {
       '10': 'value'
     },
   ],
-  '7': const {'7': true},
+  '7': {'7': true},
 };
 
 /// Descriptor for `Union`. Decode as a `google.protobuf.DescriptorProto`.
 final $typed_data.Uint8List unionDescriptor = $convert.base64Decode(
-    'CgVVbmlvbhIkCgxib29sZWFuVmFsdWUYASABKAhIAFIMYm9vbGVhblZhbHVlEiIKC3N0cmluZ1ZhbHVlGAIgASgJSABSC3N0cmluZ1ZhbHVlEjkKCWRhdGVWYWx1ZRgDIAEoCzIZLmdpdGpvdXJuYWwuRGF0ZVRpbWVBbnlUekgAUglkYXRlVmFsdWUSHAoIaW50VmFsdWUYBCABKANIAFIIaW50VmFsdWUSGAoGaXNOdWxsGAcgASgISABSBmlzTnVsbBIvCglsaXN0VmFsdWUYBSADKAsyES5naXRqb3VybmFsLlVuaW9uUglsaXN0VmFsdWUSOwoIbWFwVmFsdWUYBiADKAsyHy5naXRqb3VybmFsLlVuaW9uLk1hcFZhbHVlRW50cnlSCG1hcFZhbHVlGk4KDU1hcFZhbHVlRW50cnkSEAoDa2V5GAEgASgJUgNrZXkSJwoFdmFsdWUYAiABKAsyES5naXRqb3VybmFsLlVuaW9uUgV2YWx1ZToCOAFCDAoKVW5pb25PbmVvZg==');
+    'CgVVbmlvbhIkCgxib29sZWFuVmFsdWUYASABKAhIAFIMYm9vbGVhblZhbHVlEiIKC3N0cmluZ1'
+    'ZhbHVlGAIgASgJSABSC3N0cmluZ1ZhbHVlEjkKCWRhdGVWYWx1ZRgDIAEoCzIZLmdpdGpvdXJu'
+    'YWwuRGF0ZVRpbWVBbnlUekgAUglkYXRlVmFsdWUSHAoIaW50VmFsdWUYBCABKANIAFIIaW50Vm'
+    'FsdWUSGAoGaXNOdWxsGAcgASgISABSBmlzTnVsbBIvCglsaXN0VmFsdWUYBSADKAsyES5naXRq'
+    'b3VybmFsLlVuaW9uUglsaXN0VmFsdWUSOwoIbWFwVmFsdWUYBiADKAsyHy5naXRqb3VybmFsLl'
+    'VuaW9uLk1hcFZhbHVlRW50cnlSCG1hcFZhbHVlGk4KDU1hcFZhbHVlRW50cnkSEAoDa2V5GAEg'
+    'ASgJUgNrZXkSJwoFdmFsdWUYAiABKAsyES5naXRqb3VybmFsLlVuaW9uUgV2YWx1ZToCOAFCDA'
+    'oKVW5pb25PbmVvZg==');
+
 @$core.Deprecated('Use dateTimeAnyTzDescriptor instead')
-const DateTimeAnyTz$json = const {
+const DateTimeAnyTz$json = {
   '1': 'DateTimeAnyTz',
-  '2': const [
-    const {'1': 'timestamp', '3': 1, '4': 1, '5': 4, '10': 'timestamp'},
-    const {'1': 'offset', '3': 2, '4': 1, '5': 5, '10': 'offset'},
+  '2': [
+    {'1': 'timestamp', '3': 1, '4': 1, '5': 4, '10': 'timestamp'},
+    {'1': 'offset', '3': 2, '4': 1, '5': 5, '10': 'offset'},
   ],
 };
 
 /// Descriptor for `DateTimeAnyTz`. Decode as a `google.protobuf.DescriptorProto`.
 final $typed_data.Uint8List dateTimeAnyTzDescriptor = $convert.base64Decode(
-    'Cg1EYXRlVGltZUFueVR6EhwKCXRpbWVzdGFtcBgBIAEoBFIJdGltZXN0YW1wEhYKBm9mZnNldBgCIAEoBVIGb2Zmc2V0');
+    'Cg1EYXRlVGltZUFueVR6EhwKCXRpbWVzdGFtcBgBIAEoBFIJdGltZXN0YW1wEhYKBm9mZnNldB'
+    'gCIAEoBVIGb2Zmc2V0');
+
 @$core.Deprecated('Use noteSerializationSettingsDescriptor instead')
-const NoteSerializationSettings$json = const {
+const NoteSerializationSettings$json = {
   '1': 'NoteSerializationSettings',
-  '2': const [
-    const {'1': 'modifiedKey', '3': 1, '4': 1, '5': 9, '10': 'modifiedKey'},
-    const {'1': 'createdKey', '3': 2, '4': 1, '5': 9, '10': 'createdKey'},
-    const {'1': 'titleKey', '3': 3, '4': 1, '5': 9, '10': 'titleKey'},
-    const {'1': 'typeKey', '3': 4, '4': 1, '5': 9, '10': 'typeKey'},
-    const {'1': 'tagsKey', '3': 5, '4': 1, '5': 9, '10': 'tagsKey'},
-    const {'1': 'tagsInString', '3': 6, '4': 1, '5': 8, '10': 'tagsInString'},
-    const {'1': 'tagsHaveHash', '3': 7, '4': 1, '5': 8, '10': 'tagsHaveHash'},
-    const {'1': 'emojify', '3': 8, '4': 1, '5': 8, '10': 'emojify'},
-    const {
+  '2': [
+    {'1': 'modifiedKey', '3': 1, '4': 1, '5': 9, '10': 'modifiedKey'},
+    {'1': 'createdKey', '3': 2, '4': 1, '5': 9, '10': 'createdKey'},
+    {'1': 'titleKey', '3': 3, '4': 1, '5': 9, '10': 'titleKey'},
+    {'1': 'typeKey', '3': 4, '4': 1, '5': 9, '10': 'typeKey'},
+    {'1': 'tagsKey', '3': 5, '4': 1, '5': 9, '10': 'tagsKey'},
+    {'1': 'tagsInString', '3': 6, '4': 1, '5': 8, '10': 'tagsInString'},
+    {'1': 'tagsHaveHash', '3': 7, '4': 1, '5': 8, '10': 'tagsHaveHash'},
+    {'1': 'emojify', '3': 8, '4': 1, '5': 8, '10': 'emojify'},
+    {
       '1': 'modifiedFormat',
       '3': 9,
       '4': 1,
@@ -352,7 +376,7 @@ const NoteSerializationSettings$json = const {
       '6': '.gitjournal.DateFormat',
       '10': 'modifiedFormat'
     },
-    const {
+    {
       '1': 'createdFormat',
       '3': 10,
       '4': 1,
@@ -360,14 +384,8 @@ const NoteSerializationSettings$json = const {
       '6': '.gitjournal.DateFormat',
       '10': 'createdFormat'
     },
-    const {
-      '1': 'titleSettings',
-      '3': 11,
-      '4': 1,
-      '5': 9,
-      '10': 'titleSettings'
-    },
-    const {
+    {'1': 'titleSettings', '3': 11, '4': 1, '5': 9, '10': 'titleSettings'},
+    {
       '1': 'unixTimestampMagnitude',
       '3': 12,
       '4': 1,
@@ -379,6 +397,14 @@ const NoteSerializationSettings$json = const {
 };
 
 /// Descriptor for `NoteSerializationSettings`. Decode as a `google.protobuf.DescriptorProto`.
-final $typed_data.Uint8List noteSerializationSettingsDescriptor =
-    $convert.base64Decode(
-        'ChlOb3RlU2VyaWFsaXphdGlvblNldHRpbmdzEiAKC21vZGlmaWVkS2V5GAEgASgJUgttb2RpZmllZEtleRIeCgpjcmVhdGVkS2V5GAIgASgJUgpjcmVhdGVkS2V5EhoKCHRpdGxlS2V5GAMgASgJUgh0aXRsZUtleRIYCgd0eXBlS2V5GAQgASgJUgd0eXBlS2V5EhgKB3RhZ3NLZXkYBSABKAlSB3RhZ3NLZXkSIgoMdGFnc0luU3RyaW5nGAYgASgIUgx0YWdzSW5TdHJpbmcSIgoMdGFnc0hhdmVIYXNoGAcgASgIUgx0YWdzSGF2ZUhhc2gSGAoHZW1vamlmeRgIIAEoCFIHZW1vamlmeRI+Cg5tb2RpZmllZEZvcm1hdBgJIAEoDjIWLmdpdGpvdXJuYWwuRGF0ZUZvcm1hdFIObW9kaWZpZWRGb3JtYXQSPAoNY3JlYXRlZEZvcm1hdBgKIAEoDjIWLmdpdGpvdXJuYWwuRGF0ZUZvcm1hdFINY3JlYXRlZEZvcm1hdBIkCg10aXRsZVNldHRpbmdzGAsgASgJUg10aXRsZVNldHRpbmdzEloKFnVuaXhUaW1lc3RhbXBNYWduaXR1ZGUYDCABKA4yIi5naXRqb3VybmFsLlVuaXhUaW1lc3RhbXBNYWduaXR1ZGVSFnVuaXhUaW1lc3RhbXBNYWduaXR1ZGU=');
+final $typed_data.Uint8List noteSerializationSettingsDescriptor = $convert.base64Decode(
+    'ChlOb3RlU2VyaWFsaXphdGlvblNldHRpbmdzEiAKC21vZGlmaWVkS2V5GAEgASgJUgttb2RpZm'
+    'llZEtleRIeCgpjcmVhdGVkS2V5GAIgASgJUgpjcmVhdGVkS2V5EhoKCHRpdGxlS2V5GAMgASgJ'
+    'Ugh0aXRsZUtleRIYCgd0eXBlS2V5GAQgASgJUgd0eXBlS2V5EhgKB3RhZ3NLZXkYBSABKAlSB3'
+    'RhZ3NLZXkSIgoMdGFnc0luU3RyaW5nGAYgASgIUgx0YWdzSW5TdHJpbmcSIgoMdGFnc0hhdmVI'
+    'YXNoGAcgASgIUgx0YWdzSGF2ZUhhc2gSGAoHZW1vamlmeRgIIAEoCFIHZW1vamlmeRI+Cg5tb2'
+    'RpZmllZEZvcm1hdBgJIAEoDjIWLmdpdGpvdXJuYWwuRGF0ZUZvcm1hdFIObW9kaWZpZWRGb3Jt'
+    'YXQSPAoNY3JlYXRlZEZvcm1hdBgKIAEoDjIWLmdpdGpvdXJuYWwuRGF0ZUZvcm1hdFINY3JlYX'
+    'RlZEZvcm1hdBIkCg10aXRsZVNldHRpbmdzGAsgASgJUg10aXRsZVNldHRpbmdzEloKFnVuaXhU'
+    'aW1lc3RhbXBNYWduaXR1ZGUYDCABKA4yIi5naXRqb3VybmFsLlVuaXhUaW1lc3RhbXBNYWduaX'
+    'R1ZGVSFnVuaXhUaW1lc3RhbXBNYWduaXR1ZGU=');
