@@ -24,6 +24,7 @@ import 'package:gitjournal/core/folder/sorting_mode.dart';
 import 'package:gitjournal/core/notes/note.dart';
 import 'package:gitjournal/editors/common_types.dart';
 import 'package:gitjournal/editors/note_editor_selection_dialog.dart';
+import 'package:gitjournal/folder_listing/view/folder_listing.dart';
 import 'package:gitjournal/folder_views/common_types.dart';
 import 'package:gitjournal/folder_views/folder_view_configuration_dialog.dart';
 import 'package:gitjournal/folder_views/folder_view_selection_dialog.dart';
@@ -31,7 +32,6 @@ import 'package:gitjournal/folder_views/standard_view.dart';
 import 'package:gitjournal/iap/purchase_thankyou_screen.dart';
 import 'package:gitjournal/iap/purchase_widget.dart';
 import 'package:gitjournal/l10n.dart';
-import 'package:gitjournal/screens/folder_listing.dart';
 import 'package:gitjournal/screens/onboarding_screens.dart';
 import 'package:gitjournal/settings/settings_git_remote.dart';
 import 'package:gitjournal/settings/settings_git_widgets.dart';
